requirements.txt
alembic==1.13.1
annotated-types==0.6.0
anyio==4.3.0
blinker==1.7.0
certifi==2024.2.2
click==8.1.7
distro==1.9.0
dnspython==2.6.1
email_validator==2.1.1
Flask==3.0.2
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.6
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==5.2.1
Mako==1.3.2
MarkupSafe==2.1.5
openai==1.16.2
packaging==24.0
pluggy==1.4.0
pydantic==2.6.4
pydantic_core==2.16.3
PyMySQL==1.1.0
pytest==8.1.1
pytest-mock==3.14.0
python-docx==1.1.0
python-dotenv==1.0.1
repo2txt==0.1.6
sniffio==1.3.1
SQLAlchemy==2.0.28
tqdm==4.66.2
typing_extensions==4.10.0
Werkzeug==3.0.1
